Correções edição e exclusão

+++ Fiap.Web.Alunos/Controllers/AgendamentoController.cs
@@ -49,7 +49,7 @@ namespace Fiap.Web.Alunos.Controllers
 
         // Anotação de uso do Verb HTTP Get
         [HttpGet]
-        public IActionResult Detail(int id)
+        public IActionResult Detail(long id)
         {
             var agendamento = _context.Agendamentos.FirstOrDefault(a => a.Id == id); // Encontra o agendamento pelo id
 
@@ -65,7 +65,7 @@ namespace Fiap.Web.Alunos.Controllers
 
         // Anotação de uso do Verb HTTP Get
         [HttpGet]
-        public IActionResult Edit(int id)
+        public IActionResult Edit(long id)
         {
             var agendamento = _context.Agendamentos.Find(id);
             if (agendamento == null)
@@ -88,7 +88,7 @@ namespace Fiap.Web.Alunos.Controllers
         }
 
         [HttpGet]
-        public IActionResult Delete(int id)
+        public IActionResult Delete(long id)
         {
             var agendamento = _context.Agendamentos.Find(id);
             if (agendamento != null)
